Fix text squeezing for record type name and note fields
Basically just tweaked autolayout contraints to enable long names to stretch out or be truncated / fit properly

+++ ryCal/ryCal/Record.m
@@ -74,7 +74,7 @@
     [query whereKey:kDateFieldKey lessThan:endDate];
 
     [query findObjectsInBackgroundWithBlock:^(NSArray *objects, NSError *error) {
-        NSLog(@"Loaded %ld ENABLED records for time range: %@, %@", objects.count, startDate, endDate);
+        NSLog(@"Loaded %ld ENABLED records for time range: %@, %@", (unsigned long)objects.count, startDate, endDate);
         if (!error) {
             [[RecordQueryTracker sharedQueryTracker] updateDatastore:key objects:objects];
         }
